Add indicators

+++ benchmarks/rust/indicators/bench_macd.rs
@@ -2,6 +2,7 @@ use criterion::BenchmarkId;
 use rand::rngs::StdRng;
 use rand::{Rng, SeedableRng};
 use techalysis::indicators::macd::macd;
+use techalysis::types::Float;
 
 fn bench_macd(c: &mut criterion::Criterion) {
     let mut bench_group = c.benchmark_group("macd");
@@ -10,7 +11,7 @@ fn bench_macd(c: &mut criterion::Criterion) {
 
     for (len, fastperiod, slowperiod, signalperiod) in cases {
         let mut rng = StdRng::seed_from_u64(fastperiod as u64);
-        let data: Vec<f64> = (0..len).map(|_| rng.random_range(0.0..100.0)).collect();
+        let data: Vec<Float> = (0..len).map(|_| rng.random_range(0.0..100.0)).collect();
 
         bench_group.bench_with_input(
             BenchmarkId::new(
