+++ crates/python/src/py_ema.rs
@@ -1,21 +1,22 @@
 use numpy::{IntoPyArray, PyArray1, PyArrayMethods, PyReadonlyArray1, PyUntypedArrayMethods};
 use pyo3::{exceptions::PyValueError, pyclass, pyfunction, pymethods, Py, PyResult, Python};
 use techalysis::indicators::ema::{ema_into, ema_next as core_ema_next, EmaState};
+use techalysis::types::Float;
 
 #[derive(Debug, Clone)]
 #[pyclass(name = "EmaState", module = "techalysis._core")]
 pub struct PyEmaState {
     #[pyo3(get)]
-    pub ema: f64,
+    pub ema: Float,
     #[pyo3(get)]
     pub period: usize,
     #[pyo3(get)]
-    pub alpha: Option<f64>,
+    pub alpha: Option<Float>,
 }
 #[pymethods]
 impl PyEmaState {
     #[new]
-    pub fn new(ema: f64, period: usize, alpha: Option<f64>) -> Self {
+    pub fn new(ema: Float, period: usize, alpha: Option<Float>) -> Self {
         PyEmaState { ema, period, alpha }
     }
     #[getter]
@@ -43,11 +44,11 @@ impl From<EmaState> for PyEmaState {
 #[pyfunction(signature = (data, period = 14, alpha = None, release_gil = false))]
 pub(crate) fn ema(
     py: Python,
-    data: PyReadonlyArray1<f64>,
+    data: PyReadonlyArray1<Float>,
     period: usize,
-    alpha: Option<f64>,
+    alpha: Option<Float>,
     release_gil: bool,
-) -> PyResult<(Py<PyArray1<f64>>, PyEmaState)> {
+) -> PyResult<(Py<PyArray1<Float>>, PyEmaState)> {
     let len = data.len();
     let input_slice = data.as_slice()?;
 
@@ -58,7 +59,7 @@ pub(crate) fn ema(
             .map_err(|e| PyValueError::new_err(format!("{:?}", e)))?;
         Ok((output.into_pyarray(py).into(), state.into()))
     } else {
-        let output_array = PyArray1::<f64>::zeros(py, [len], false);
+        let output_array = PyArray1::<Float>::zeros(py, [len], false);
         let output_slice = unsafe { output_array.as_slice_mut()? };
         let state = ema_into(input_slice, period, alpha, output_slice)
             .map_err(|e| PyValueError::new_err(format!("{:?}", e)))?;
@@ -67,7 +68,7 @@ pub(crate) fn ema(
 }
 
 #[pyfunction(signature = (new_value, ema_state))]
-pub(crate) fn ema_next(new_value: f64, ema_state: PyEmaState) -> PyResult<PyEmaState> {
+pub(crate) fn ema_next(new_value: Float, ema_state: PyEmaState) -> PyResult<PyEmaState> {
     let state = core_ema_next(new_value, ema_state.ema, ema_state.period, ema_state.alpha)
         .map_err(|e| PyValueError::new_err(format!("{:?}", e)))?;
     Ok(state.into())
