+++ tests/fuzz/fuzz_targets/fuzz_ema.rs
@@ -2,8 +2,9 @@
 
 use libfuzzer_sys::fuzz_target;
 use techalysis::indicators::ema::ema;
+use techalysis::types::Float;
 
-fuzz_target!(|data: (Vec<f64>, u8, f64)| {
+fuzz_target!(|data: (Vec<Float>, u8, Float)| {
     let (v, w, s) = data;
     let w = (w as usize % v.len().saturating_add(1)).max(1);
     let _ = ema(&v, w, s.into());
