+++ tests/fuzz/fuzz_targets/fuzz_sma.rs
@@ -2,8 +2,9 @@
 
 use libfuzzer_sys::fuzz_target;
 use techalysis::indicators::sma::sma;
+use techalysis::types::Float;
 
-fuzz_target!(|data: (Vec<f64>, u8)| {
+fuzz_target!(|data: (Vec<Float>, u8)| {
     let (v, w) = data;
     let w = (w as usize % v.len().saturating_add(1)).max(1);
     let _ = sma(&v, w);
